Rotated init

+++ Projekt2/src/shapes/Location.java
@@ -11,10 +11,10 @@ package shapes;
  */
 public class Location {
 
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Location(int x, int y) {
+    public Location(float x, float y) {
         this.x = x;
         this.y = y;
     }
@@ -22,19 +22,19 @@ public class Location {
     public Location() {
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
